requirements.txt
python-telegram-bot>=21.0
anthropic>=0.40.0
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.1
google-api-python-client>=2.0.0
rapidfuzz>=3.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pyobjc-framework-Vision>=10.0
pyobjc-framework-Quartz>=10.0
